# artifacts/api-server/src/lib/approval-service.ts
/**
 * @file        artifacts/api-server/src/lib/approval-service.ts
 * @module      API Server / Services / Approval
 * @purpose     Business logic for approval workflow operations
 *
 * @ai_instructions
 *   - Contains all approval-related business logic
 *   - Uses database through the db instance from lib/db
 *   - Handles approval decisions with transaction safety
 *   - Updates agent status based on approval decisions
 *   - Follows repository pattern for data access
 *   - Includes proper error handling and tenant isolation
 *
 * @exports     ApprovalService class and singleton instance
 * @imports     @workspace/db, drizzle-orm
 *
 * @copyright   SPDX-FileCopyrightText: 2025 Trevor Lam <trevor@example.org>
 * @license     SPDX-License-Identifier: MIT
 */

import { 
  desc, 
  eq, 
  and
} from "drizzle-orm";
import { db } from "@workspace/db";
import { 
  approvalRequestsTable,
  approvalDecisionsTable,
  agentsTable,
  type ApprovalRequest,
  type ApprovalDecision
} from "@workspace/db/schema";
import { ErrorTypes } from "../middlewares/error-handler";

interface ListApprovalsParams {
  page: number;
  limit: number;
  status: string;
  tenantId: string;
}

interface DecideApprovalParams {
  approvalRequestId: string;
  decision: "approved" | "rejected";
  comment?: string;
  decidedBy: string;
  decidedByName: string;
  tenantId: string;
}

interface PaginationResult {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}

interface ApprovalListResponse {
  approvals: (ApprovalRequest & { agentName?: string })[];
  pagination: PaginationResult;
}

interface ApprovalDecisionResponse {
  decision: ApprovalDecision;
  approvalRequest: ApprovalRequest;
}

/**
 * Approval service containing business logic for approval operations
 */
export class ApprovalService {
  /**
   * List approval requests with pagination and status filtering
   */
  async listApprovals(params: ListApprovalsParams): Promise<ApprovalListResponse> {
    const { page, limit, status, tenantId } = params;
    const offset = (page - 1) * limit;

    // Build base query with tenant isolation and status filter
    const baseQuery = and(
      eq(approvalRequestsTable.tenantId, tenantId),
      eq(approvalRequestsTable.status, status as any)
    ) as any;

    // Get total count for pagination
    const countResult = await db
      .select({ count: approvalRequestsTable.id })
      .from(approvalRequestsTable)
      .where(baseQuery);
    
    const total = countResult.length;

    // Get paginated results with agent information
    const approvals = await db
      .select({
        id: approvalRequestsTable.id,
        agentRunId: approvalRequestsTable.agentRunId,
        tenantId: approvalRequestsTable.tenantId,
        title: approvalRequestsTable.title,
        description: approvalRequestsTable.description,
        status: approvalRequestsTable.status,
        priority: approvalRequestsTable.priority,
        requestContext: approvalRequestsTable.requestContext,
        requiredBy: approvalRequestsTable.requiredBy,
        expiresAt: approvalRequestsTable.expiresAt,
        createdAt: approvalRequestsTable.createdAt,
        updatedAt: approvalRequestsTable.updatedAt,
        resolvedAt: approvalRequestsTable.resolvedAt
      })
      .from(approvalRequestsTable)
      .where(baseQuery)
      .orderBy(desc(approvalRequestsTable.createdAt))
      .limit(limit)
      .offset(offset);

    // TODO: Join with agent_runs and agents to get agent name
    // For now, return approvals without agent name

    return {
      approvals,
      pagination: {
        page,
        limit,
        total,
        totalPages: Math.ceil(total / limit)
      }
    };
  }

  /**
   * Record a decision on an approval request and update related entities
   */
  async decideApproval(params: DecideApprovalParams): Promise<ApprovalDecisionResponse> {
    const { 
      approvalRequestId, 
      decision, 
      comment, 
      decidedBy, 
      decidedByName, 
      tenantId 
    } = params;

    // Use a transaction to ensure data consistency
    return await db.transaction(async (tx) => {
      // Get the approval request
      const approvalRequests = await tx
        .select()
        .from(approvalRequestsTable)
        .where(and(
          eq(approvalRequestsTable.id, approvalRequestId!),
          eq(approvalRequestsTable.tenantId, tenantId)
        ) as any)
        .limit(1);

      if (approvalRequests.length === 0) {
        throw ErrorTypes.NotFound("Approval request");
      }

      const approvalRequest = approvalRequests[0];

      // Check if the approval is still pending
      if (approvalRequest.status !== "pending") {
        throw ErrorTypes.Conflict("Approval request has already been decided");
      }

      // Create the approval decision
      const decisionResult = await tx
        .insert(approvalDecisionsTable)
        .values({
          approvalRequestId,
          decision,
          decidedBy,
          decidedByName,
          comment,
          decidedAt: new Date().toISOString(),
          tenantId
        })
        .returning();

      const decisionRecord = decisionResult[0];

      // Update the approval request status and resolved timestamp
      await tx
        .update(approvalRequestsTable)
        .set({
          status: decision,
          resolvedAt: new Date().toISOString(),
          updatedAt: new Date().toISOString()
        })
        .where(eq(approvalRequestsTable.id, approvalRequestId!));

      // TODO: Update the agent status based on the decision
      // This would require joining through agent_runs to get the agent_id
      // For now, we'll skip the agent status update as it's not critical for the API

      return {
        decision: decisionRecord,
        approvalRequest: {
          ...approvalRequest,
          status: decision,
          resolvedAt: new Date().toISOString(),
          updatedAt: new Date().toISOString()
        }
      };
    });
  }
}

/**
 * Singleton instance of the approval service
 */
export const approvalService = new ApprovalService();