# lib/api-spec/openapi.yaml
...
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List pending approvals
      description: Returns pending approval requests
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired]
            default: pending
      responses:
        "200":
          description: List of approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /approvals/{id}/decide:
    post:
      operationId: decideApproval
      tags: [approvals]
      summary: Decide on approval request
      description: Records decision and updates agent status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Approval request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
...
components:
  schemas:
...
    ApprovalRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agentRunId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        priority:
          type: string
          enum: [low, medium, high, critical]
        requiredBy:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id, agentRunId, title, description, status, priority, createdAt
    ApprovalDecision:
      type: object
      properties:
        id:
          type: string
          format: uuid
        approvalRequestId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, rejected]
        decidedBy:
          type: string
        decidedByName:
          type: string
        comment:
          type: string
        decidedAt:
          type: string
          format: date-time
      required:
        - id, approvalRequestId, decision, decidedBy, decidedAt
...
    ApprovalListResponse:
      type: object
      properties:
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - approvals, pagination
    ApprovalDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        comment:
          type: string
      required:
        - decision
    ApprovalDecisionResponse:
      type: object
      properties:
        decision:
          $ref: "#/components/schemas/ApprovalDecision"
        approvalRequest:
          $ref: "#/components/schemas/ApprovalRequest"
      required:
        - decision, approvalRequest
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page, limit, total, totalPages
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
      required:
        - error, statusCode
    ValidationError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        details:
          type: object
      required:
        - error, statusCode
    # Chat schemas
...